+++ record_ios/ios/record_ios/Sources/record_ios/RecordConfig.swift
@@ -81,7 +81,6 @@ public class Device {
 
 struct IosConfig {
   let categoryOptions: [AVAudioSession.CategoryOptions]
-  let manageAudioSession: Bool
   let allowHapticsAndSystemSoundsDuringRecording: Bool
 
   init(map: [String: Any]) {
@@ -114,7 +113,6 @@ struct IosConfig {
       }
     }
     self.categoryOptions = options ?? []
-    self.manageAudioSession = map["manageAudioSession"] as? Bool ?? true
     self.allowHapticsAndSystemSoundsDuringRecording = map["allowHapticsAndSystemSoundsDuringRecording"] as? Bool ?? false
   }
 }