+++ record_platform_interface/lib/src/types/ios_record_config.dart
@@ -5,13 +5,6 @@ class IosRecordConfig {
   /// https://developer.apple.com/documentation/avfaudio/avaudiosession/categoryoptions
   final List<IosAudioCategoryOption> categoryOptions;
 
-  /// Manage the shared AVAudioSession (defaults to `true`).
-  ///
-  /// Set this to `false` if another plugin is already managing the AVAudioSession.
-  ///
-  /// If `false`, there will be no check of the [RecordConfig] given.
-  final bool manageAudioSession;
-
   /// Whether haptics and system sounds (including incoming call ring tones) are
   /// allowed to play while recording (defaults to `false`).
   ///
@@ -29,14 +22,11 @@ class IosRecordConfig {
       IosAudioCategoryOption.allowBluetooth,
       IosAudioCategoryOption.allowBluetoothA2DP,
     ],
-    @Deprecated('Use AudioRecorder.ios.manageAudioSession instead.')
-    this.manageAudioSession = true,
     this.allowHapticsAndSystemSoundsDuringRecording = false,
   });
   Map<String, dynamic> toMap() {
     return {
       "categoryOptions": categoryOptions.map((e) => e.name).join(','),
-      "manageAudioSession": manageAudioSession,
       "allowHapticsAndSystemSoundsDuringRecording":
           allowHapticsAndSystemSoundsDuringRecording,
     };
@@ -62,5 +52,5 @@ enum IosAudioCategoryOption {
   allowAirPlay,
 
   /// available from iOS 14.5
-  overrideMutedMicrophoneInterruption
+  overrideMutedMicrophoneInterruption,
 }